Extract download validation helpers into bitnet-download-core
