+++ crates/bitnet-download-core/Cargo.toml
@@ -1,16 +1,16 @@
 [package]
 name = "bitnet-download-core"
-version = "0.2.1-dev"
-edition = "2024"
-description = "Core download validation and file helpers for BitNet"
-license = "MIT OR Apache-2.0"
-repository = "https://github.com/BitNet-Project/BitNet"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+documentation.workspace = true
+description = "SRP core download helpers for offline mode and content-length validation"
 
 [dependencies]
-thiserror = { workspace = true }
+thiserror.workspace = true
 
 [lints]
 workspace = true
-
-[dev-dependencies]
-tempfile = { workspace = true }
